backend/requirements.txt
fastapi
uvicorn
psycopg[binary]
python-dotenv
pytest
